Updated PCB (v1.2)
Added a flexPCB connector on the edge of the board connected to the ICSP pins, making easier to program the ATMEGA328P.
The battery terminals footprint is now only on one side of the board, preventing accidental short-circuit on the bottom side.
Moved some components to allow more space for the LiPo battery.
Changed the values of components C7,C9 R5,R6, adjusting the values for filtering the audio output.